Remove global function closures and unnecessary tests
