+++ ospatch/updates_linux.go
@@ -24,7 +24,6 @@ import (
 	"time"
 
 	"github.com/GoogleCloudPlatform/osconfig/inventory/packages"
-	"github.com/GoogleCloudPlatform/osconfig/util"
 
 	osconfigpb "github.com/GoogleCloudPlatform/osconfig/_internal/gapi-cloud-osconfig-go/google.golang.org/genproto/googleapis/cloud/osconfig/v1alpha2"
 )
@@ -43,7 +42,7 @@ func (r *patchRun) systemRebootRequired() (bool, error) {
 		r.debugf("/var/run/reboot-required exists indicating a reboot is required, content:\n%s", string(data))
 		return true, nil
 	}
-	if ok := util.Exists(rpmquery); ok {
+	if ok := exists(rpmquery); ok {
 		r.debugf("Checking if reboot required by querying rpm database.")
 		return rpmReboot()
 	}
