+++ util/util.go
@@ -21,7 +21,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -150,28 +149,6 @@ func newHTTPClient() *http.Client {
 		return testHTTPClient
 	}
 	return &http.Client{}
-}
-
-// Stubbed methods below
-// this is done so that this function can be stubbed
-// for unit testing
-
-// Exists Checks if a file exists on the filesystem
-var Exists = func(name string) bool {
-	if _, err := os.Stat(name); os.IsNotExist(err) {
-		return false
-	}
-	return true
-}
-
-// OsHostname is a wrapper to get os hostname
-var OsHostname = func() (name string, err error) {
-	return os.Hostname()
-}
-
-// ReadFile is a wrapper to read file
-var ReadFile = func(file string) ([]byte, error) {
-	return ioutil.ReadFile(file)
 }
 
 // Run is a wrapper to execute terminal commands
